isolate class gen demo
