+++ InputAndOutput.java
@@ -2,7 +2,9 @@ import java.io.IOException;
 
 public class InputAndOutput {
   public static void main(String[] args) throws IOException {
-    int input = System.in.read();   // reads ASCII code
-    System.out.println((char) input);
+    byte[] memory = new byte[30000];
+    int head = 0;
+    memory[head] = (byte) System.in.read();   // reads ASCII code
+    System.out.println((char) memory[head]);
   }
 }
