+++ class_generator.js
@@ -306,17 +306,6 @@ export class ClassFileGenerator {
     this.writeU2(3); // max_locals
     this.writeU4(jvmInstructions.length); // code_length
     this.writeBytes(jvmInstructions);
-
-    // Main bytecode: getstatic System.out, ldc "Hello, World!", invokevirtual println, return
-    // this.writeU1(0xB2); // getstatic
-    // this.writeU2(outFieldrefIndex);
-    // this.writeU1(0x12); // ldc
-    // this.writeU1(helloStringConstantIndex);
-    // this.writeU1(0xB6); // invokevirtual
-    // this.writeU2(printlnMethodrefIndex);
-
-    // this.writeU1(0xB1); // return
-
     this.writeU2(0); // exception_table_length
     this.writeU2(0); // attributes_count
 
@@ -326,103 +315,7 @@ export class ClassFileGenerator {
     return new Uint8Array(this.buffer);
   }
 
-  // Utility method to save class file (for Node.js environments)
-  saveToFile(classData, filename) {
-    if (typeof require !== 'undefined') {
-      const fs = require('fs');
-      fs.writeFileSync(filename, classData);
-      console.log(`Class file saved as ${filename}`);
-    } else {
-      console.log('File saving not available in browser environment');
-      console.log('Class data:', Array.from(classData).map(b => '0x' + b.toString(16).padStart(2, '0')).join(' '));
-    }
+  static saveToFile(filename, classData) {
+    fs.writeFileSync(filename, classData);
   }
 }
-
-function intTo2Bytes(num) {
-  return [(num >> 8) & 0xFF, num & 0xFF];
-}
-
-const move_head = (h) => {
-  return []
-    .concat([0x11, ...intTo2Bytes(h)]) // sipush 10
-    .concat([0x3d]) // istore_2 (storing in head)
-    ;
-};
-
-const increment = (n) => {
-  return []
-    .concat([0x2b]) // aload_1
-    .concat([0x1c]) // iload_2
-    .concat([0x5c]) // dup2
-    .concat([0x33]) // baload
-    .concat([0x11, ...intTo2Bytes(n)]) // sipush n
-    .concat([0x60]) // iadd
-    .concat([0x91]) // i2b
-    .concat([0x54]) // bastore
-    ;
-};
-
-const input = ({ fieldRef, methodRef }) => {
-  const getStatic = [0xb2, ...intTo2Bytes(fieldRef)];
-  const invokeVirtual = [0xb6, ...intTo2Bytes(methodRef)];
-  return [
-    ...getStatic,
-    ...invokeVirtual,
-    0x2b, // aload_1 (load arrayref)
-    0x5f, // swap
-    0x1c, // iload_2 (load index)
-    0x5f, // swap
-    0x54, // baload
-  ]
-};
-
-// stack: arrayref, index, value
-
-const output = ({ fieldRef, methodRef }) => {
-  const getStatic = [0xb2, ...intTo2Bytes(fieldRef)];
-  const invokeVirtual = [0xb6, ...intTo2Bytes(methodRef)];
-  return [
-    ...getStatic,
-    ...[
-      0x2b, // aload_1
-      0x1c, // iload_2
-      0x33, // baload (get value at index -> cell[head])
-      0x92, // i2c
-    ], 
-    ...invokeVirtual
-  ]
-};
-
-const makeJVMInstructions = ({
-  constantPool,
-  symbolicConstantPool
-}) => {
-  const readInput = Buffer.from(input({
-    fieldRef: symbolicConstantPool.input.fieldRef, 
-    methodRef: symbolicConstantPool.input.readMethodrefIndex
-  }));
-
-  const printCharToStdOut = Buffer.from(output({
-    fieldRef: symbolicConstantPool.output.fieldRef, 
-    methodRef: symbolicConstantPool.output.printlnMethodrefIndex
-  }))
-
-  return [
-    ...[0x11, ...intTo2Bytes(30_000)], // sipush 30000
-    ...[0xbc, 0x08], // newarray byte
-    ...[0x4c], // astore_1 (cells)
-    ...[0x03], // iconst_0
-    ...[0x3d], // istore_2 (head)
-    ...readInput,
-    ...printCharToStdOut,
-    // ...increment(100),
-    0xb1 // return
-  ];
-};
-
-const className = process.argv[2] || 'HelloWorld';
-const generator = new ClassFileGenerator();
-const helloWorldClass = generator.generateHelloWorldClass(className, makeJVMInstructions);
-console.log(`${className}.class generated:`, helloWorldClass.length, 'bytes');
-fs.writeFileSync(`${className}.class`, helloWorldClass);
